Refactored recording code

+++ app/src/main/kotlin/io/github/leonidius20/recorder/data/recordings_list/RecordingsDataSourceAndroid10And11.kt
@@ -75,7 +75,7 @@ class RecordingsDataSourceAndroid10And11 @Inject constructor(
                 // Get values of columns for a given video.
                 val id = cursor.getLong(idColumn)
                 val name = cursor.getString(nameColumn)
-                val duration = cursor.getInt(durationColumn)
+                val duration = cursor.getLong(durationColumn)
                 val size = cursor.getInt(sizeColumn)
                 val dateTaken = cursor.getLong(dateTakenColumn)
                 val mimeType = cursor.getString(mimeTypeColumn)
